SkiaSharp mode

+++ CSJ2K/j2k/codestream/Markers.cs
@@ -9,10 +9,10 @@
 *
 * COPYRIGHT:
 * 
-* This software module was originally developed by Raphaël Grosbois and
+* This software module was originally developed by Raphaï¿½l Grosbois and
 * Diego Santa Cruz (Swiss Federal Institute of Technology-EPFL); Joel
-* Askelöf (Ericsson Radio Systems AB); and Bertrand Berthelot, David
-* Bouchard, Félix Henry, Gerard Mozelle and Patrice Onno (Canon Research
+* Askelï¿½f (Ericsson Radio Systems AB); and Bertrand Berthelot, David
+* Bouchard, Fï¿½lix Henry, Gerard Mozelle and Patrice Onno (Canon Research
 * Centre France S.A) in the course of development of the JPEG2000
 * standard as specified by ISO/IEC 15444 (JPEG 2000 Standard). This
 * software module is an implementation of a part of the JPEG 2000
@@ -152,6 +152,6 @@ namespace CSJ2K.j2k.codestream
 		/// <summary>Comment (COM): 0xFF64 </summary>
 		public const short COM = unchecked((short)0xff64);
 		/// <summary>General use registration value (COM): 0x0001 </summary>
-		public const short RCOM_GEN_USE = unchecked((short)0x0001);
+		public const short RCOM_GEN_USE = unchecked(0x0001);
 	}
 }
